refactor: move envcheck and mutter class
