+++ desktop/src/cn/harryh/arkpets/guitasks/envchecker/WinGraphicsEnvCheckTask.java
@@ -1,4 +1,4 @@
-package cn.harryh.arkpets.envchecker;
+package cn.harryh.arkpets.guitasks.envchecker;
 
 import cn.harryh.arkpets.utils.IOUtils;
 import cn.harryh.arkpets.utils.Logger;
